% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{create_hierarchy}
\alias{create_hierarchy}
\title{Create hierarchical tree}
\usage{
create_hierarchy(branches)
}
\arguments{
\item{branches}{A character vector containing \code{n} number of branches.
Each branch corresponds to a cell type. Each level is delimited by  "/"
character. For example, the following branch is composed by three levels
and the _terminal_ cell type is \code{pDC}: "Myeloid/DC/pDC"}
}
\value{
A data.tree object containing the cell type hierarchy
}
\description{
Creates a tree object representing a cell-type hierarchy
}
\examples{

branches <- c("Myeloid/DC/pDC", "Myeloid/DC/cDC")
h <- create_hierarchy(branches)

}
\author{
Lieke Michielsen and Jose Alquicira-Hernandez
}
