% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate performance of hierarchical classifier}
\usage{
evaluate(y_true, y_pred, tree)
}
\arguments{
\item{y_true}{vector with true labels}

\item{y_pred}{vector with predicted labels}

\item{tree}{data.tree object storing the hierarchy used during classification}
}
\value{
A list containing the different evaluation metrics (median HF1-score, HF1-score per cell population,
median F1-score, F1-score per cell population, % rejected cells, % internally labeled cells)
}
\description{
Calculated hierarchical F1-score, median F1-score, percentage of rejected cells, 
percentage of internally labeled cells, confusion matrix, and population size
}
\examples{

library(data.tree)

res <- evaluate(y_true, y_pred, tree)

}
\author{
Lieke Michelsen and Jose Alquicira-Hernandez
}
