% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{findSet}
\alias{findSet}
\title{Find ancestor of a node in the tree}
\usage{
findSet(name, tree)
}
\arguments{
\item{name}{name of the node}

\item{tree}{data.tree object storing the hierarchy used during classification}
}
\value{
A set containing the name of the node and its ancestors
}
\description{
Return a set containing the node itself and its ancestors
}
\examples{

library(data.tree)

nodeSet <- findSet('B cells', tree)

}
\author{
Lieke Michielsen and Jose Alquicira-Hernandez
}
