% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predictNode}
\alias{predictNode}
\title{Predict labels of a dataset}
\usage{
predictNode(tree, newData, threshold, recompute_alignment)
}
\arguments{
\item{tree}{data.tree object storing the hierarchy and trained classifiers per layer}

\item{newData}{Seurat object containing cells to annotate}

\item{threshold}{Threshold used for probabilities to classify cells into classes. All cells below 
this threshold value will be labels as "unassigned".}

\item{recompute_alignment}{Recompute alignment?}
}
\value{
A Seurat object with additional metadata columns with prediction probabilities associated to each 
class, a \code{prediction} column, indicating the classification based on the provided threshold and 
a \code{generic_class} column without "unassigned" labels.
}
\description{
Recursive function to predict the labels of a dataset using a tree object
}
\examples{

}
\author{
Lieke Michielsen and Jose Alquicira-Hernandez
}
