% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predictTree}
\alias{predictTree}
\title{Predict labels of a dataset}
\usage{
predictTree(tree, newData, threshold = 0, max.iter.harmony = 10)
}
\arguments{
\item{tree}{data.tree object storing the hierarchy and trained classifiers per layer}

\item{newData}{Seurat object containing cells to annotate}

\item{threshold}{Threshold used for probabilities to classify cells into classes. All cells below
this threshold value will be labels as "unassigned".}

\item{max.iter.harmony}{Maximum number of clustering iterations by harmony}
}
\value{
A Seurat object with additional metadata columns with prediction probabilities associated to each
class, a \code{prediction} column, indicating the classification based on the provided threshold and
a \code{generic_class} column without "unassigned" labels.
}
\description{
Predict the labels of a dataset using a tree object
}
\examples{

branches <- c("Myeloid/DC/pDC", "Myeloid/DC/cDC")
h <- create_hierarchy(branches)
train_tree(dataTrain, h)
dataTest <- predictTree(h, dataTest)
}
\author{
Lieke Michielsen and Jose Alquicira-Hernandez
}
