% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{train_node}
\alias{train_node}
\title{Create hierarchical tree}
\usage{
train_node(
  tree,
  data,
  pvar,
  reduction,
  model,
  reconstruction_error,
  fn_perc,
  verbose
)
}
\arguments{
\item{tree}{Hierarchy created via \code{create_hierarchy}}

\item{data}{Seurat object containing cells used to train hierarchical models}

\item{pvar}{Prediction variable. Column name in Seurat object metadata
containing the cell labels of the terminal nodes of the hierarchy tree}

\item{reduction}{Name of reduction in Seurat objet to be used to determine
the feature space. Default: "pca"}

\item{model}{Classification model supported via caret package.
A list of all models can be found here in https://topepo.github.io/caret/available-models.html}

\item{reconstruction_error}{Use reconstruction error?}

\item{fn_perc}{TBD}
}
\value{
A data.tree object containing the cell type hierarchy
}
\description{
Creates a tree object representing a cell-type hierarchy
}
\examples{

branches <- c("Myeloid/DC/pDC", "Myeloid/DC/cDC")
h <- create_hierarchy(branches)
train_tree(data, h)

}
\author{
Lieke Michelsen and Jose Alquicira-Hernandez
}
